Custom node name, node opts, and init opts
This adds a way to provide the node name and node options, as well as a
way to externally manage init/shutdown on a context. If the given node
options have a valid context, then DrakeRos won't call init or shutdown
on it. In this way external code can pass init options.

Signed-off-by: Shane Loretz <sloretz@openrobotics.org>
Signed-off-by: Shane Loretz <sloretz@osrfoundation.org>